Fatmug_VideoCC/requirements.txt
Django==3.2.16
webvtt-py==0.5.1
psycopg2==2.9.9
